Changed windows parameter to a class, added class mapper to database, refactoring database (not finished), refined the evaluation and their schema

+++ src/es/unileon/happycow/model/composite2/EvaluationCategory.java
@@ -17,7 +17,7 @@ public class EvaluationCategory extends Composite {
         this(idHandler, 1);
     }
 
-    public EvaluationCategory(IdHandler idHandler, int weighing) {
+    public EvaluationCategory(IdHandler idHandler, float weighing) {
         super();
         this.entity = Entity.CATEGORY;
         id = idHandler;
